refactor(repository, evaluation): evaluation repository 통합
