+++ src/main/java/com/pnu/momeet/domain/evaluation/repository/EvaluationRepository.java
@@ -2,7 +2,6 @@ package com.pnu.momeet.domain.evaluation.repository;
 
 import com.pnu.momeet.domain.evaluation.entity.Evaluation;
 import java.time.LocalDateTime;
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -10,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EvaluationRepository extends JpaRepository<Evaluation, UUID> {
+public interface EvaluationRepository extends JpaRepository<Evaluation, UUID>, EvaluationDslRepository {
 
     boolean existsByMeetupIdAndEvaluatorProfileIdAndTargetProfileId(
         UUID meetupId,
